[utils] Rename json_writer.* to json_utils.*.
