+++ include/multipass/json_utils.h
@@ -17,19 +17,14 @@
  *
  */
 
-#include <multipass/file_ops.h>
-#include <multipass/json_writer.h>
+#ifndef MULTIPASS_JSON_UTILS_H
+#define MULTIPASS_JSON_UTILS_H
 
-#include <QFile>
-#include <QJsonDocument>
+#include <QJsonObject>
+#include <QString>
 
-namespace mp = multipass;
-
-void mp::write_json(const QJsonObject& root, QString file_name)
+namespace multipass
 {
-    QJsonDocument doc{root};
-    auto raw_json = doc.toJson();
-    QFile db_file{file_name};
-    MP_FILEOPS.open(db_file, QIODevice::ReadWrite | QIODevice::Truncate);
-    MP_FILEOPS.write(db_file, raw_json);
+void write_json(const QJsonObject& root, QString file_name);
 }
+#endif // MULTIPASS_JSON_UTILS_H
