+++ src/utils/CMakeLists.txt
@@ -16,7 +16,7 @@ function(add_target TARGET_NAME)
   add_library(${TARGET_NAME} STATIC
     file_ops.cpp
     memory_size.cpp
-    json_writer.cpp
+    json_utils.cpp
     snap_utils.cpp
     standard_paths.cpp
     timer.cpp
